removed DomainCollection::fromValue(), GenericPaginatedDomainCollection
