+++ src/Domain/Infrastructure/Console/Definition/ClassContextDefinition.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace MsgPhp\Domain\Infrastructure\Console\Definition;
 
-use MsgPhp\Domain\DomainCollection;
 use MsgPhp\Domain\DomainId;
 use MsgPhp\Domain\Factory\ClassMethodResolver;
 use MsgPhp\Domain\Infrastructure\Console\Context\ClassContextElementFactory;
@@ -245,7 +244,7 @@ final class ClassContextDefinition implements DomainContextDefinition
      */
     private function resolveNested(string $class, array $parentValue, ContextElement $parentElement): array
     {
-        $method = is_subclass_of($class, DomainCollection::class) || is_subclass_of($class, DomainId::class) ? 'fromValue' : '__construct';
+        $method = is_subclass_of($class, DomainId::class) ? 'fromValue' : '__construct';
         $resolved = [];
 
         foreach (ClassMethodResolver::resolve($class, $method) as $argument => $metadata) {
